fix(upstream): reorganises String exercises
- converts skeleton from standalone class to sub-package application
- improves README instructions
