--- exercises/java-primitives/src/main/java/com/cbfacademy/StringExercises.java
@@ -1,19 +1,0 @@
-package com.cbfacademy;
-
-public class StringExercises {
-
-        throw new RuntimeException("Not implemented");
-    public String concatenate(String word1, String word2, String word3) {
-        // TODO: Write code that concatenates the input parameters and returns the result
-    }
-
-        throw new RuntimeException("Not implemented");
-    public Boolean areEqual(String word1, String word2) {
-        // TODO: Write code to determine whether the input parameters are equal strings
-    }
-
-        throw new RuntimeException("Not implemented");
-    public String format(String item, int quantity, double price) {
-        // TODO: Write code to return a string formatted as follows: "Item: [item]. Price: £[amount]. Quantity: [quantity]". The date should be formatted as DD/MM/YYYY, e.g. 24/10/2023
-    }
-}
